one more step to better code
- but still far away
